requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.0
requests>=2.31.0
